Added "Chaos Archfiend Nehemoth"
[Pendulum Effect]
You can Special Summon this card (from your Pendulum Zone) by banishing 1 LIGHT and 1 DARK monster from your GY. If this card is Special Summoned from your Pendulum Zone it gains this effect.
●When a card or effect is activated (Quick Effect): You can banish 1 card from your GY; negate the activation, and if you do, destroy it. You can only use this effect of "Chaos Archfiend Nehemoth" once per turn.
---------------------------------
[Monster Effect]
[1 LIGHT Tuner + 1+ LIGHT non-Tuner OR 1 DARK non-Tuner - 1 Dark Tuner]
(This card is also Level -8)
If this card is Synchro Summoned using a LIGHT monster as material it gains these effect.
●This card is treated as a LIGHT monster while face-up on the field.
●You can send 1 DARK monster from your Deck to the GY; Special Summon 1 LIGHT monster from your hand or GY.
If this card is Synchro Summoned using a DARK monster as material it gains this effect.
●You can banish 1 LIGHT monster from your GY; add 1 DARK monster with the same Level from your Deck to your hand, then take 500 damage. You can only use the above effects of "Chaos Archfiend Nehemoth" once per turn. If this card is destroyed: You can place this card in your Pendulum Zone.