Added "Papillon the Emperror Butterfly"
1 Tuner + 1+ non-Tuner monsters
If this card is Synchro Summoned: You can target 1 Insect monster in your GY; add it to your hand, and if you do, inflict damage to both players equal to the ATK of that monster. If this card is sent to the GY for a Synchro Summon: You can target 1 Level 4 or lower monster in your GY; Inflict damage to your opponent equal to the ATK of the target. You can only use each effect of "Papillon the Emperror Butterfly" once per turn.