Added "Berserk Bear"
If a card(s) on the field is destroyed: You can Special Summon this card from the GY (if it was there when the card was destroyed) or hand (even if not). You can only use this effect of "Berserk Bear" once per turn. While this card is face-up on the field, non-Effect Synchro Monsters you control cannot be destroyed by battle. This card gains 200 ATK for each non-Effect Synchro Monster on the field. If this card is destroyed: You can target 1 non-Effect Synchro Monster you control; equip this card to it. A monster equipped with this card gains following effect.
●If this card destroys a monster by battle: You can add 1 Level 4 or lower EARTH Beast monster from your Deck to your hand.