changed hex of SUMMON_TYPE_TRICK
was overlapping with pendulum summon for some reason, i dont know if i fixed it for good so please try the trick monsters. if something happens that shouldnt write me on discord: coroln

known bugs: when you have the right materials but no linked zone to summon too then it will still be shown that you can summon, but when you try nothing will happen. so you can say it is just visual.

sorry for all the bugs en annoyances with the trick monster but the last update of edopro killed it and i cannot do anything against it (not their fault they didnt know)