added "Attack Commander, Lodrian" (Trick)
1 Effect Monster + 1 Trap
If this card is Trick Summoned: You can move this card to another Main Monster Zone, then, Place 4 counters on it. When a monster declares an attack: You can remove 1 counter from this card; negate the attack, and if you do, increase this cards ATK by 300. An Xyz Monster that was Summoned using this card as Xyz Material gains this effect:
● This card can make a second attack during each Battle Phase.

[ Artwork by: krysfun - https://www.deviantart.com/krysfun ]