Added some cards
"Pottery Archfiend"
You can Special Summon this card (from your hand) by discarding 1 Spell that includes an effect that adds a card(s) from the Deck to the hand. If Summoned this way: draw 1 card. This card gains 100 ATK for each card in your hand. If this card destroys a monster by battle: You can activate 1 of these effects;
●Draw 2 cards, then discard 1 card.
●Destroy up to 2 Spell/Traps your opponent controls.
●Banish up to 2 cards from your opponent's GY.
During the End Phase, if this card was Special Summoned this turn: Return this card to your hand, and if you do, draw 1 card, then discard 1 card.
---------------------------------
"Fluff Synchron"
If you control a Tuner Monster or a Synchro Monster, you can Special Summon this card (from your hand). If this card is used as a Synchro material, you can treat it as a non-Tuner. Once per turn: You can declare a Level from 1 to 7; this card becomes that Level until the end of this turn.