Added "Infinite Dark Storm"
[Field Trap]
You can activate this card when a monster is Summoned. Each time a monster is Normal or Special Summoned, except from the Extra Deck or from underneath this card: place that monster(s) underneath this card. Once per turn: The turn player can Special Summon 1 monster that is placed underneath this card to their side of the field. If this card in the field Zone is destroyed: banish all monsters that are placed underneath this card, and if you do, both player take damage equal to the banish cards x300.